Stacks in Java
Stack is an abstract data structure that follows the LIFO rule, which means Last In First Out. In this data structure, the Last item is inserted first, and at the time of removing, the first item is removed. The best real-life example for this is the stack of books, the stake of plates, etc.